User functionality/roles/authentication

+++ src/app/inventory/inventory.component.ts
@@ -1,8 +1,12 @@
 import { Component, OnInit, ViewChild, Inject } from '@angular/core';
-import { MatSort, MatTableDataSource, MatDialog, MatDialogRef, MatPaginator, MatSlideToggleChange } from '@angular/material';
+import { MatSort, MatTableDataSource, MatDialog, MatDialogRef, MatPaginator, MatSlideToggleChange, MatCheckbox } from '@angular/material';
 import { AdditemComponent } from '../additem/additem.component';
 import { InventoryService } from './inventory.service';
 import { Item } from './inventory.model';
+import { User } from '../shared/services/user.model';
+import { SnackbarService } from '../shared/snackbar.service';
+import { AngularFireDatabase } from 'angularfire2/database';
+import { AngularFireAuth } from 'angularfire2/auth';
 
 
 @Component({
@@ -17,6 +21,14 @@ export class InventoryComponent implements OnInit {
 
   itemList: Item[];
   selectedItem: Item;
+ 
+  user: User;
+
+  public tf: boolean;
+
+  usertf: boolean;
+  admintf: boolean;
+  superusertf: boolean;
 
   //edit data
   edit = new Item("","","","","",true,true,true);
@@ -37,8 +49,13 @@ export class InventoryComponent implements OnInit {
     this.dataSource.filter = filterValue.trim().toLowerCase();
   }
 
-  constructor(public dialog: MatDialog, private inventoryService: InventoryService) {
+  constructor(public dialog: MatDialog,
+    private inventoryService: InventoryService,
+    private snack: SnackbarService,
+    private db: AngularFireDatabase,
+    private afAuth: AngularFireAuth) {
     inventoryService.getAllItems();
+
   }
 
 
@@ -66,6 +83,7 @@ export class InventoryComponent implements OnInit {
 
   //initializing data
   ngOnInit() {
+
     let data = this.inventoryService.getAllItems();
     data.snapshotChanges().subscribe(item => {
       this.itemList = [];
@@ -82,10 +100,23 @@ export class InventoryComponent implements OnInit {
   }
 
 
+
   //Delete/Edit functions
   onDelete(key: string) {
-    this.inventoryService.deletebyKey(key);
-    this.refreshAfterEdit();
+
+    this.db.database.ref('/users/').child(this.afAuth.auth.currentUser.uid).child('role')
+    .child('superuser').once('value').then(snapshot => {
+      this.tf = snapshot.val();
+      console.log(this.tf);
+    });
+
+    if (this.tf) {
+      this.inventoryService.deletebyKey(key);
+
+    }
+    else if (!!this.tf) {
+      this.snack.openSnackBar('You do not have permission for this', 2000);
+    }
   }
   
   selectItem(key: string, modal: string) {
@@ -99,7 +130,16 @@ export class InventoryComponent implements OnInit {
   }
 
   selectJoinedToggle(key: string) {
+    this.db.database.ref('/users/').child(this.afAuth.auth.currentUser.uid).child('role')
+    .child('admin').once('value').then(snapshot => {
+      this.tf = snapshot.val();
+      console.log(this.tf);
+    })
+
+    if (this.tf === true) {
+
     this.selectedItem = this.itemList.filter(x => x.$key === key)[0];
+
     if (this.selectedItem.joined === true) {
       this.edit.joined = false;
       let editedItem = new Item(this.selectedItem.mac, this.selectedItem.location, this.selectedItem.port, this.selectedItem.created_at,
@@ -115,11 +155,23 @@ export class InventoryComponent implements OnInit {
       editedItem.lastUpdate = new Date().toString();
   
       this.inventoryService.editItem(this.selectedItem.$key, editedItem);
-    }
+    };
+  }
+  else if (this.tf === false) {
+    this.snack.openSnackBar('You do not have permission for this.', 2000);
+  };
     this.refreshAfterEdit();
   }
 
   selectCompleteToggle(key: string) {
+    this.db.database.ref('/users/').child(this.afAuth.auth.currentUser.uid).child('role')
+    .child('admin').once('value').then(snapshot => {
+      this.tf = snapshot.val();
+      console.log(this.tf);
+    })
+
+    if (this.tf === true) {
+
     this.selectedItem = this.itemList.filter(x => x.$key === key)[0];
     if (this.selectedItem.complete === true) {
       this.edit.complete = false;
@@ -136,11 +188,23 @@ export class InventoryComponent implements OnInit {
       editedItem.lastUpdate = new Date().toString();
   
       this.inventoryService.editItem(this.selectedItem.$key, editedItem);
-    }
+    };
+  }
+  else if (this.tf === false) {
+    this.snack.openSnackBar('You do not have permission for this.', 2000);
+  };
     this.refreshAfterEdit();
   }
 
   selectCheckedInToggle(key: string) {
+    this.db.database.ref('/users/').child(this.afAuth.auth.currentUser.uid).child('role')
+    .child('admin').once('value').then(snapshot => {
+      this.tf = snapshot.val();
+      console.log(this.tf);
+    })
+
+    if (this.tf === true) {
+
     this.selectedItem = this.itemList.filter(x => x.$key === key)[0];
 
     if (this.selectedItem.checkedIn === true) {
@@ -154,16 +218,31 @@ export class InventoryComponent implements OnInit {
         this.selectedItem.created_by, this.selectedItem.joined, this.selectedItem.complete, true);
       editedItem.lastUpdate = new Date().toString();
       this.inventoryService.editItem(this.selectedItem.$key, editedItem);
-    }
+    };
+  }
+    else if (this.tf === false) {
+      this.snack.openSnackBar('You do not have permission for this.', 2000);
+    };
     this.refreshAfterEdit();
   }
 
   onSave() {
+    this.db.database.ref('/users/').child(this.afAuth.auth.currentUser.uid).child('role')
+    .child('admin').once('value').then(snapshot => {
+      this.tf = snapshot.val();
+      console.log(this.tf);
+    })
+
+    if (this.tf === true) {
     let editedItem = new Item(this.edit.mac, this.edit.location, this.edit.port, this.selectedItem.created_at,
       this.selectedItem.created_by, this.selectedItem.joined, this.selectedItem.complete, this.selectedItem.checkedIn);
     editedItem.lastUpdate = new Date().toString();
 
     this.inventoryService.editItem(this.selectedItem.$key, editedItem);
+    }
+    else if (this.tf === false) {
+      this.snack.openSnackBar('You do not have permission for this.', 2000);
+    };
     this.refreshAfterEdit();
   }
